requirements.txt
Django==1.9.4
django-mptt==0.8.3
